rz-cmn-srp: alsa-utils: restore and prefer version 1.2.11 from Scarthgap branch
Due to a regression in alsa-utils 1.2.12 causing silent playback during
audio capture on the RZ/V2H-EVK, we restore version 1.2.11 from the
Scarthgap branch and pin it as the preferred version.

The issue stems from a change in WAV header generation for the S24_LE
format (24-bit audio in a 32-bit container), which breaks compatibility
with certain audio drivers.

Signed-off-by: Long Trinh <long.trinh.rh@bp.renesas.com>
Signed-off-by: Tien Nguyen <tien.nguyen.uh@renesas.com>